Latency now int

+++ src/dataset_code_score.py
@@ -7,7 +7,7 @@ from api_keys import open_ai_key
 
 ERROR_VALUE = -1.0
 
-async def compute(model_url: str, code_url: str | None, dataset_url: str | None) -> Tuple[float, float]:
+async def compute(model_url: str, code_url: str | None, dataset_url: str | None) -> Tuple[float, int]:
     """
     Grade how well a Hugging Face model's code and dataset are documented.
     
@@ -24,7 +24,7 @@ async def compute(model_url: str, code_url: str | None, dataset_url: str | None)
         model_id = model_url.replace("https://huggingface.co/", "").strip("/") #removing the huggingface part of the link to get the model id
     except Exception as e:
         print(f"Error parsing Hugging Face link: {e}") #exception if huggingface is wrong
-        return ERROR_VALUE, (time.time() - start_time) * 1000
+        return ERROR_VALUE, (int)((time.time() - start_time) * 1000)
 
     # --- Step 2: Fetch model card & datasets ---
     try:
@@ -37,7 +37,7 @@ async def compute(model_url: str, code_url: str | None, dataset_url: str | None)
         datasets = data.get("datasets", [])
     except Exception as e:
         print(f"Error fetching model info: {e}")
-        return ERROR_VALUE, (time.time() - start_time) * 1000
+        return ERROR_VALUE, (int)((time.time() - start_time) * 1000)
 
     # Handle missing dataset
     dataset_text = "\n".join(datasets) if datasets else "NO DATASET PROVIDED" #if no datasets provided, return no dataset provided
@@ -82,7 +82,7 @@ async def compute(model_url: str, code_url: str | None, dataset_url: str | None)
 {joined}"""
 
     if not combined_text.strip():  # if no text at all, return error
-        return ERROR_VALUE, (time.time() - start_time) * 1000
+        return ERROR_VALUE, (int)((time.time() - start_time) * 1000)
 
     # --- Step 5: Prompt LLM for grading ---
     prompt = f"""
@@ -113,7 +113,7 @@ Repository content:
         )
         score_text = response["choices"][0]["message"]["content"].strip()
         score = float(score_text)
-        return max(0.0, min(1.0, score)), (time.time() - start_time) * 1000  # clamp to [0,1]
+        return max(0.0, min(1.0, score)), (int)((time.time() - start_time) * 1000)  # clamp to [0,1]
     except Exception as e:
         print(f"Error grading documentation: {e}")
-        return ERROR_VALUE, (time.time() - start_time) * 1000
+        return ERROR_VALUE, (int)((time.time() - start_time) * 1000)
